requirements.txt
requests==2.23.0
terminaltables==3.1.0
python-dotenv==0.13.0
argparse==1.4.0